Disable 1.4 build, avoid the gowrappers rebuild.
All builds using gowrappers fail until they are rebuilt due to some part
of the build refusing to proceed if the go tool timestamp is newer than
part of the library that is being used in the build. The gowrappers use
golang 1.4-swig, which is out of support and not getting updates.
Disable building it so that we can build 1.5 and 1.6 without breaking
everything using 1.4.